% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CircularDDM.R
\name{bessel_zero}
\alias{bessel_zero}
\title{Find First k Zeros of a Bessel Function}
\usage{
bessel_zero(nu, k, kind)
}
\arguments{
\item{nu}{The order of the corresponding Bessel function.}

\item{k}{an integer for the first k positive zeros.}

\item{kind}{0, 1, or 2. A switch selects \link{besselI}, \link{besselJ} or
\link{besselY}}
}
\value{
a vector
}
\description{
The function modifies from Von Winckel's \code{besselzero.m}, using Halley's
method to find the first k zeros of a Bessel function.
The Bessel function is either the first kind, \code{R::bessel_j} or the
second kind, \code{R::bessel_y}.
}
\examples{
nu <- seq(0, 5, length.out=10)
output <- matrix(numeric(5*length(nu)), nrow=5)
for (i in 1:length(nu)) { output[,i] <- bessel_zero(nu[i], 5, 1) }
print(output)

output <- matrix(numeric(5*length(nu)), nrow=5)
for(i in 1:length(nu)) { output[,i] <- bessel_zero(nu[i], 5, 2) }
print(output)
}
