% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CircularDDM.R
\docType{package}
\name{CircularDDM}
\alias{CircularDDM}
\alias{CircularDDM-package}
\title{Circular Drift-diffusion Model}
\description{
\code{CircularDDM} provides density functions for circular
decision-diffusion/decision-diffusion model for continuous report. It
implements equations and convenient functions
for calculating likelihood and generating random deviates. For circular
DDM, random deviates are continuous response times and response angles.
}
\references{
\enumerate{
      \item Smith, P. L. (2016). Diffusion Theory of Decision Making in
      Continuous Report, \emph{Psychological Review}, 123 (4), 425--451.
      \url{http://dx.doi.org/10.1037/rev0000023}.
      \item Hamana, Y., & Matsumoto, H. (2013). The probability distributions
      of the first hitting times of Bessel processes, \emph{Transactions}
      \emph{of the American Mathematical Society}, 365, 5237--5257.
      \item Agostinelli, C. and Lund, U (2017). R package 'circular':
      Circular Statistics (version 0.4-93). \url{https://r-forge.r-project.org/projects/circular/}.
      \item Von Winckel, G. besselzero.m
      \url{http://au.mathworks.com/matlabcentral/fileexchange/6794-bessel-function-zeros/content/besselzero.m}
}
}
\author{
Yi-Shin Lin <yishin.lin@utas.edu.au> \cr
Andrew Heathcote <andrew.heathcote@utas.edu.au> \cr
Peter Kvam <kvam.peter@gmail.com> \cr
}
\keyword{CircularDDM}
