% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CircularDDM.R
\name{conv_drift}
\alias{conv_drift}
\title{Convert Dirft Rate from Magnitude-Direction to X-Y coordinate or the other way}
\usage{
conv_drift(mean_v, mag2xy)
}
\arguments{
\item{mean_v}{a \code{c(magnitude, direction)} or \code{c(mu_x mu_y)} vector}

\item{mag2xy}{a boolean switch indicating converting from \code{c(magnitude,
direction)} (TRUE) to \code{c(mu_x, mu_y)} or the other way (FALSE).}
}
\value{
a vector
}
\description{
Convert dirft rate from magnitude-direction to X-Y coordinate or the other
way around.
}
\examples{
conv_drift(c(.5, -pi/3), TRUE)
conv_drift(c(.25, -0.4330127), FALSE)

}
