% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcddm1}
\alias{rcddm1}
\alias{rcddm2_internal}
\title{The Circular Decision-diffusion Distribution}
\usage{
rcddm1(n, a, mu1, mu2, t0, sigma, stepSize = 1, stepTime = 1e-04)

rcddm2_internal(n, threshold, angle, sp, t0, stepSize = 0.1,
  stepTime = 0.005)
}
\arguments{
\item{n}{numbers of observation.}

\item{a}{decision threshold. Must be a scalar.}

\item{mu1}{x axis drift rate. This must be a scalar.}

\item{mu2}{y axis drift rate. This must must be a scalar.}

\item{t0}{non-decision time. This must be a scalar.}

\item{sigma}{square root of the diffusion coefficient. This is the
within-trial, moment-to-moment, standard deviation. This must be a scalar.}

\item{stepSize}{the size of a random walk step.}

\item{stepTime}{the unit time for a random walk step.}

\item{threshold}{a threshold vector.}

\item{angle}{an angle vector, allowing one to supply a vector of randomly
drawn numbers from any distributions.}

\item{sp}{a starting point matrix.  First column must be the x coordinate
of the starting point and the second column must be the y coordinate of
the starting point. Each row is a random number.}
}
\value{
a n x 2 matrix with two columns: RTs and angles.
}
\description{
Generate random deviates from a circular decision-diffusion
model, using parameters, a, vx, vy, t0, and s. \code{rcddm1} is
canonical form of the random number generator for the circular
decision-diffusion model. \code{rcddm2} is an extension, allowing one to
supply a threshold vector, an angle vector, a starting point matrix
[xPos, yPos] and a non-decision time. The angle vector carries the
information of the drift magnitude and the drift direction.
}
\examples{
## rcddm1 example
den  <- rcddm1(1024, a = 2, mu1 = 1.5, mu2 = 1.25, t0 = 0.25, sigma = 1);
par(mfrow = c(1, 2))
hist(den[,1], breaks = "fd", xlab="RT (s)",  main="")
hist(den[,2], breaks = "fd", xlab="Angle (Rad)", main="")

## rcddm2 example
t0 <- .25
threshold <- runif(10, 0, 2)
angle <- rvm(1e3, 0, 10)
startpoint <- cbind(runif(10, 0, 1), runif(10, 0,1))
den  <- rcddm2_internal(1e3, threshold, angle, startpoint, t0)
head(den)

## If starting points only has one row, this is equilvalent to fixing
## the starting points at constant values. This is because internally the
## function will always use the first row when 'sp' is a row vector.
## Similarly, one can apply the same trick on threshold and angle vectors.
startpoint <- cbind(runif(1, 0, 1),  runif(1, 0,1));
den <- rcddm2_internal(1e3, threshold, angle, startpoint, t0)

## If the user enters a R's vector, which is not a matrix,
## rcddm2 wrapper helps the user to convert it to a matrix (ie row vector)
## Note this flexibility requires checking steps.
den <- rcddm2_internal(1e3, threshold, angle, startpoint, t0)

}
\references{
Smith, P. L. (2016). Diffusion Theory of Decision Making in
Continuous Report, Psychological Review, 123 (4), 425--451.
}
