% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rvm}
\alias{rvm}
\title{Von Mises Distribution: Random Number Generator}
\usage{
rvm(n, mu, k)
}
\arguments{
\item{n}{number of observations.}

\item{mu}{mean direction of the distribution.}

\item{k}{non-negative numeric value for the concentration parameter of the
distribution}
}
\value{
a vector
}
\description{
Generate random numbers for von Mises circular distribution.
}
\examples{
n  <- 100
mu <- 0
k  <- 10
vm3_de <- rvm(n, mu, k)       ## in degree unit
vm3_pi <- vm3_de \%\% (2 * pi)  ## in radian unit
}

