% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CircularDDM.R
\docType{package}
\name{CircularDDM}
\alias{CircularDDM}
\alias{CircularDDM-package}
\title{Circular Drift-diffusion Model}
\description{
Circular drift-diffusion model for continuous report.
}
\references{
Smith, P. L. (2016). Diffusion Theory of Decision Making in
Continuous Report, Psychological Review, 123(4), 425--451.
}
\author{
Yi-Shin Lin <yishin.lin@utas.edu.au> \cr
Andrew Heathcote <andrew.heathcote@utas.edu.au> \cr
Peter Kvam <kvam.peter@gmail.com> \cr
}
\keyword{CircularDDM}
