% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CircularDDM.R, R/RcppExports.R
\name{rcddm2}
\alias{rcddm2}
\alias{rcddm1}
\alias{rcddm2_internal}
\title{The Circular Drift-diffusion Distribution}
\usage{
rcddm2(n, threshold, angle, sp, t0, p = 0.15, tol = 1000L)

rcddm1(n, pVec, p = 0.15)

rcddm2_internal(n, threshold, angle, sp, t0, p = 0.15, tol = 1000L)
}
\arguments{
\item{n}{number of observations.}

\item{threshold}{a threshold vector.}

\item{angle}{an angle vector, allowing one to supply a vector of random
draws from an arbitary, instead of von Mise, distribution.}

\item{sp}{a starting point matrix.  First column must be the x coordinate
of the starting point and the second column must be the y coordinate of
the starting point. Each row is a random deviate.}

\item{t0}{non-decision time. Must be a scalar.}

\item{p}{a precision for random walk step in \code{rcddm}. Default is 0.15
second}

\item{tol}{an upper bound (step) for the drift-diffusion process. This
prevents an unrealistic threshold been enterd from falling into while loop
trap. Default is 1,000 steps.}

\item{pVec}{a parameter vector with the order [a, vx, vy, t0, sv_sq].
a stands for response threshold, vx is the drift rate along x axis,
vy is the drift rate along y axis, t0 is the non-decision time, and sv_sq
is the within-trial variance. The order matters.}
}
\value{
a n x 2 matrix with two columns: RTs and angles.
}
\description{
Generate random deviates for the circular drift-diffusion
model with a theta vector, namely \code{pVec}. \code{rcddm1} is a canonical
form of the random number generator for circular drift-diffusion model.
\code{rcddm2} is an extension, allowing one to supply a threshold vector,
an angle vector, a starting point matrix [xPos, yPos] and a non-decision
time.
}
\examples{
## rcddm1 example
pVec <- c(a=2, vx=1.5, vy=1.25, t0=.25, s=1)
den  <- rcddm1(1e3, pVec);
hist(den[,1], breaks = "fd", xlab="Response Time",  main="Density")
hist(den[,2], breaks = "fd", xlab="Response Angle", main="Density")

## rcddm2 example
t0 <- .25
threshold <- runif(10, 0, 2)
angle <- rvm(1e3, 0, 10)
startpoint <- cbind(runif(10, 0, 1), runif(10, 0,1))
den  <- rcddm2_internal(1e3, threshold, angle, startpoint, t0)
head(den)

## If starting points only has one row, this is equilvalent to fixing
## the starting points at constant values. This is because internally the
## function will always use the first row when 'sp' is a row vector.
## Similarly, one can apply the same trick on threshold and angle vectors.
startpoint <- cbind(runif(1, 0, 1),  runif(1, 0,1));
den  <- rcddm2(1e3, threshold, angle, startpoint, t0)

## If the user enters a R's vector, which is not a matrix,
## rcddm2 wrapper helps the user to convert it to a matrix (ie row vector)
## Note this flexibility requires checking steps.
den  <- rcddm2(1e3, threshold, angle, startpoint, t0)

}
\references{
Smith, P. L. (2016). Diffusion Theory of Decision Making in
Continuous Report, Psychological Review, 123 (4), 425--451.
}
