% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rvm}
\alias{rvm}
\title{Generate random deviates for the von Mises distribution}
\usage{
rvm(n, mu, k)
}
\arguments{
\item{n}{number of observations.}

\item{mu}{mean direction of the distribution.}

\item{k}{non-negative numeric value for the concentration parameter of the
distribution}
}
\value{
a vector
}
\description{
Generate random deviates fro the von Mises distribution.
}
\details{
A random variable for circular normal distribution has the form:\cr
\deqn{f(theta; mu, kappa) = 1 / (2*pi*I_0(kappa)) * exp(kappa * cos(theta-mu))}
0 <= theta <= 2*pi

I_0(kappa) in the normalizing constant is the modified Bessel function of
the first kind and order zero. It is given by:\cr
\deqn{I_0(kappa) = 1/2*pi $$\int_{0}^{2*pi} exp(k*cos(theta)) dtheta$$}
}
\examples{
n  <- 100
mu <- 0
k  <- 10
vm3_de <- rvm(n, mu, k)       ## in degree unit
vm3_pi <- vm3_de \%\% (2 * pi)  ## in radian unit
}

